SY-4338: Export Node, drop parent cast, merge aether property test
